Added Timer to Game
Added a Timer to the game's clock
Starts off at one minute, and each goal adds 10 seconds to the clock. If the timer runs out, the blob with the most points wins. Otherwise, the game is a tie.

Added placeholder icons for kicks, boosts and blocks.

Fixed a minor CSS scaling glitch.