Incorporated Time Limit and Time Bonus
Time Limit and Bonus can be changed in the Rules Menu

Time Limit defaults to 3600 (1 minute) and can be changed in increments of 10 seconds. It can range from 0 (No limit) to 36000 (10 minutes)

Time bonus defaults to 600 (10 seconds) and can be changed in increments of 5 seconds. It can range from 0 (No bonus) to 3600 (1 minute).

These rules will impact the time limit in game.