Added many new features!
Added a Heart Icon image to the files.
Updated the kick_icon.png file to show only the boot, since it was too small before,

Added the skeleton for the UI, this is a rough idea of how I want the game to display this critical info.

Updated Water Blob's ability: Ball Y speed is now capped at -25 so the ball doesn't shoot up to the ceiling for 200 NRG.