Optimizations and Crash Fix
Optimized Gameplay to make it not load in the Blob Special image every frame... twice.

Ensured that all images are scaled once at the start of the game, instead of scaling each image every frame

Fixed a crash bug that was caused by a tie, which then confused the program since it had no idea what to write to the matchup chart.

Optimized CSS image loading, and ensured that scaling was cached instead of performed each frame.