Updated Damage Dealing Function
blob.take_damage() is a function now, and is what is used to deal damage to blobs.

Added damage_flash.png and draw_damage_flash() under display_particles.py to create a flash when a blob takes damage.

 blob.recharge_indicators gained 'damage' and 'damage_flash' keys.

Changed the way kick damage is calculated, but didn't change the end result.