Updated CPU Behavior
CPU Memory is now stored a little differently: it is now a dict with a few keys rather than a list. These fields are press_queue (which remembers buttons to press on the next frame), game_state (which remembers the last frame's game state such as opening, offense, etc) and current play (which is the strategy being used at the moment).

CPU Logic has been updated! There's a lot to go over so I'll sum up the changes here
- The game checks the position of the blobs and the ball, which are generally relative. These positions are home_dz, home_mid, mid, away_mid, and away_dz. This helps to demarcate where blobs are relative to each other

- The game determines if the CPU is winning or not for use in decision making

- Handle_Logic_Beta is the current function being used for CPU activities.

- Blobs can start with 3 openings: Charge energy (but are still unable to use it!), Rush the ball to kick it, or Rush the Ball and wait.

- The standard CPU code has been updated. CPUs will now kick the ball using their kick button if it is nearby. CPUs will now not wavebounce like crazy.

- 'toward' and 'away' are valid commands that the CPU can respond to.