Further Optimizations to Display Gameplay
UI Icons are now drawn onto a general surface at the start of the game, which is then drawn every frame. This is faster then drawing each image individually.

Optimized draw_blob_special so it uses cached versions of the special shells - this is faster than loading and then updating the special shells.

Updated the blob_drawing function so it now uses the cached left/right versions. If blobs are invisible, they are just not drawn. This is faster than reloading the blob images every time they take damage, or change direction.

P2 Blob darkening now only occurs at the start of the game.