Fixed Medal Screen (Slightly)
Removed selector from Game Handler - the selector should only be managed locally in the medal milestone file
Note for Sunken: We try to make each of the navigators the same, and each selector should go to only one navigation file to prevent random issues involving overwriting.

Updated medal_navigation() parameters - it now only has the timer parameter. This is in line with other navigators of its type.
Note for Sunken: In general, we don't need to pass the player parameter. Instead, we call engine.handle_input.merge_inputs() on line 31 to merge p1 and p2 inputs into one input.

Updated mouse to work regardless of player.

Return statement now returns game_state, and an array which is the info_getter. This is done in two parts for a specific reason: we want to access game_state quickly and intuitively, and the info_getter contains miscellaneous information that could vary in size. As such, game_display will take in the game_state (to determine the "who" of drawing) and info_getter (to determine the "what" of drawing)