Improved Controller Support
Updated the way that controller binding works. The game now properly swaps player ports and boots players in an intuitive way.

Fixed a crash bug related to pressing DPad Down (GC Controller).

Fixed "Enter" not working to start a game in the CSS.

Added a placeholder entry in the game_handler for the controller config screen