Added Popup displays for each valid event
Added a popup screen for each popup event (-1 through 8, and 10+). These are near finalized.

Centralized text in the popups, will readjust as necessary.

Updated ControllerPopUp class to take the controller name as an input. Updated Nodes to have a timer of 120 frames (up from 60)

Added preliminary XBox360 mapping, but you can't bind this yet.

Moved binding of GCCA outside of the get_keypress function to help prevent crashes (for now)