Pre Compilation Fixes
Generic controllers are now copies of XBox controllers - most controllers are going to be read this way.

Updated Version # from v0.14.0a to v0.14.0b - we are nearing release!

Prevented crashes on the main menu

UI Mode now defaults to bottom for new players. This setting does not get reset when pressing that button.