Progress on Costume Unlocks
Added Generic Popup - event ID goes in and it outputs a specific thing. Specifically, there is now a pop up for trying to load in a replay from the wrong version.

Began work on costume unlocks. There's a lot going on and it's killing me, honestly

Updated the actual unlock requirements for blobs - it now is both visually and mechanically accurate

Added grayscale variants to all blobs in the code