Initial Implementation of Glue Blob
Added in Glue Blob. Glue Blob is meant to be a disruptor class blob who is slow and a little squishy, but shines in denying control of the field to blobit's opponents.

Added Glue Blob to the CSS and made blobit unlockable, and there is the grayscale costume as well!

Added in environmental hazards - these are mostly WIP and have no impact on the gameplay yet. The only currently spawnable ones are glue_shot and glue_puddle. Also added glue_shot and glue_puddle particles (WIP)

Fixed a bug related to transferring an old costume save to a newer version