Spire + Thunderbolt Environmental Hazards
Spire and Thunderbolt now are classified as environmental hazards. Some code is intentionally left in in case bugs are present but otherwise it seems stable. Hitboxes have been updated to match (more or less)

Ball now has a "zapped" status effect - this draws static particles and causes the ball to stick to the ground.

Blobs now have an "all_blobs" dictionary which points to all blobs that currently exist.