Added the End Screen
Added the end screen for the tutorial. Split times are also recorded

Almanac now keeps track of your fastest clear and # of clears of the tutorial!

create_time_string now has an optional kwarg "smallest" which is the smallest unit of separation (useful for parsing out milliseconds or seconds as the smallest unit)