Various Bug Fixes for V17.1b
Updated Game Version to V17.1b

Fixed a bug where exiting out of a game while a KO animation plays causes the next game to continue playing that animation

Fixed a bug where exiting out of the game after scoring a goal causes the ball to glow green at the start of the next game.

Fixed a bug that caused Royal Loan to fade out.

Fixed a bug that caused Sharp Shadow to not stun.