Updated Bubble and Arcade Blobs
Bubble Blob's bubble now moves in a circle. When the bubble pops, the ball gets launched.

Arcade Blob's cartridges got buffed. When the enemy is in contact with the cartridge, they have hypothermia and judged applied for 60 frames

Teleporting out of a cartridge causes a kick to happen (similar to Explosive Focus)