Player Menu Toggles
Menus now toggle between player, cpu and none.

Cursors, Menus and Tokens are now drawn in separate """layers""" (not really, but sort of!) so they should only overlap in certain ways now.

CPUs are now selectable from the menu. We've regained most, if not all, of the functionality of the old menu