Blobs Rewrite Initial Commit
Patch is nonfunctional! Will crash upon starting a match.

Began rewriting blobs.py - there is now a dedicated folder for blobs. The new blobs.py should handle everything related to the blob, from initialization, to movement, to drawing itself. old_blobs.py is kept as a reference since it contains all the code of previous blobs in a very bloated mess...

Added a Blob Handler - this is meant to import all blobs and place them in an easy to use container. It is unknown if the container actually works or not.

Miscellaneous files have been changed slightly (mostly involving import paths) so the game can actually start